chore(pagination):refactor-pagination-code
- Adds paginate helper class
- Refactors  room_resources and room schemas to use the paginate helper class

[Delivers #159654368]
